feat: remove customer need admin role

+++ workflow/src/main/java/org/daming/hoteler/workflow/api/CustomerController.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.daming.hoteler.common.domain.Role;
 import org.daming.hoteler.common.response.CommonResponse;
 import org.daming.hoteler.common.response.DataResponse;
 import org.daming.hoteler.common.response.ListResponse;
@@ -11,6 +12,7 @@ import org.daming.hoteler.workflow.pojo.Customer;
 import org.daming.hoteler.workflow.pojo.request.CreateCustomerRequest;
 import org.daming.hoteler.workflow.pojo.request.UpdateCustomerRequest;
 import org.daming.hoteler.workflow.service.ICustomerService;
+import org.daming.hoteler.workflow.service.feign.IUserFeignService;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -22,6 +24,7 @@ import org.springframework.web.bind.annotation.*;
 public class CustomerController {
 
     private final ICustomerService customerService;
+    private final IUserFeignService userFeignService;
 
     @Operation(
             summary = "创建客户信息", security = { @SecurityRequirement(name = "bearer-key") },
@@ -73,15 +76,29 @@ public class CustomerController {
 
 
     @DeleteMapping("customer/{id}")
-    public CommonResponse delete(@PathVariable String id) {
+    public CommonResponse delete(@PathVariable String id, @RequestHeader(value = "Authorization") String accessToken) {
+        var userResponse = this.userFeignService.getUserByAccessToken(accessToken);
+        System.out.println("userResponse");
+        System.out.println(userResponse);
+        if (userResponse == null || userResponse.getStatus() != CommonResponse.SUCCESS_STATUS) {
+           throw new RuntimeException("非法操作");
+        }
+        var user = userResponse.getData();
+        if (user == null) {
+            throw new RuntimeException("非法操作");
+        }
+        if (user.getRoles().stream().map(Role::getName).noneMatch(r -> r.equalsIgnoreCase("admin"))) {
+            throw new RuntimeException("非法操作，只有管理员才有权限");
+        }
         this.customerService.delete(id);
-
         return new CommonResponse();
     }
 
 
-    public CustomerController(ICustomerService customerService) {
+    public CustomerController(ICustomerService customerService, IUserFeignService userFeignService) {
         super();
         this.customerService = customerService;
+        this.userFeignService = userFeignService;
     }
+
 }
